# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Study Planner API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Profile {
  id: string;
  supabaseId?: string | null;
  name: string;
  email: string;
  college?: string | null;
  createdAt?: string | null;
  updatedAt?: string | null;
}

export interface UpsertProfileRequest {
  userId: string;
  supabaseId?: string;
  name: string;
  email: string;
  college?: string;
}

export interface Subject {
  code: string;
  name: string;
  hours: number;
  difficulty: string;
  description: string;
  selected?: boolean | null;
}

export interface StudySettings {
  startTime: string;
  endTime: string;
  targetHours: number;
  includeWeekends: boolean;
  targetDate: string;
}

export interface StudyConfig {
  id: string;
  userId: string;
  courseType: string;
  branch: string;
  year: string;
  semester: string;
  subjects?: Subject[];
  studySettings?: StudySettings;
  createdAt?: string | null;
  updatedAt?: string | null;
}

export interface UpsertStudyConfigRequest {
  userId: string;
  courseType: string;
  branch: string;
  year: string;
  semester: string;
  subjects?: Subject[];
  studySettings?: StudySettings;
}

export interface QuizResult {
  id: string;
  userId: string;
  score: number;
  total: number;
  percentage: string;
  subject?: string | null;
  createdAt?: string | null;
}

export interface SaveQuizResultRequest {
  userId: string;
  score: number;
  total: number;
  percentage: number;
  subject?: string;
}

export interface StudySession {
  id: string;
  userId: string;
  subjectName: string;
  subjectCode?: string | null;
  hours: string;
  sessionDate: string;
  completed?: boolean | null;
  topic?: string | null;
  createdAt?: string | null;
}

export interface SaveStudySessionRequest {
  userId: string;
  subjectName: string;
  subjectCode?: string;
  hours: number;
  sessionDate: string;
  completed?: boolean;
  topic?: string;
}

export type GetProfileParams = {
  userId: string;
};

export type GetStudyConfigParams = {
  userId: string;
};

export type GetQuizResultsParams = {
  userId: string;
};

export type GetStudySessionsParams = {
  userId: string;
};