# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Smart Study Planner API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: profile
    description: User profile operations
  - name: config
    description: Study configuration operations
  - name: quiz
    description: Quiz result operations
  - name: sessions
    description: Study session operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: upsertProfile
      tags: [profile]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProfileRequest"
      responses:
        "200":
          description: Profile upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  /config:
    get:
      operationId: getStudyConfig
      tags: [config]
      summary: Get study configuration for a user
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Config found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyConfig"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: upsertStudyConfig
      tags: [config]
      summary: Create or update study configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertStudyConfigRequest"
      responses:
        "200":
          description: Config upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyConfig"

  /quiz:
    get:
      operationId: getQuizResults
      tags: [quiz]
      summary: Get quiz results for a user
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quiz results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizResult"
    post:
      operationId: saveQuizResult
      tags: [quiz]
      summary: Save a quiz result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveQuizResultRequest"
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"

  /sessions:
    get:
      operationId: getStudySessions
      tags: [sessions]
      summary: Get study sessions for a user
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Study sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudySession"
    post:
      operationId: saveStudySession
      tags: [sessions]
      summary: Save a study session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveStudySessionRequest"
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySession"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Profile:
      type: object
      properties:
        id:
          type: string
        supabaseId:
          type: string
          nullable: true
        name:
          type: string
        email:
          type: string
        college:
          type: string
          nullable: true
        createdAt:
          type: string
          nullable: true
        updatedAt:
          type: string
          nullable: true
      required:
        - id
        - name
        - email

    UpsertProfileRequest:
      type: object
      properties:
        userId:
          type: string
        supabaseId:
          type: string
        name:
          type: string
        email:
          type: string
        college:
          type: string
      required:
        - userId
        - name
        - email

    StudyConfig:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        courseType:
          type: string
        branch:
          type: string
        year:
          type: string
        semester:
          type: string
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/Subject"
        studySettings:
          $ref: "#/components/schemas/StudySettings"
        createdAt:
          type: string
          nullable: true
        updatedAt:
          type: string
          nullable: true
      required:
        - id
        - userId
        - courseType
        - branch
        - year
        - semester

    UpsertStudyConfigRequest:
      type: object
      properties:
        userId:
          type: string
        courseType:
          type: string
        branch:
          type: string
        year:
          type: string
        semester:
          type: string
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/Subject"
        studySettings:
          $ref: "#/components/schemas/StudySettings"
      required:
        - userId
        - courseType
        - branch
        - year
        - semester

    Subject:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        hours:
          type: number
        difficulty:
          type: string
        description:
          type: string
        selected:
          type: boolean
          nullable: true
      required:
        - code
        - name
        - hours
        - difficulty
        - description

    StudySettings:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
        targetHours:
          type: number
        includeWeekends:
          type: boolean
        targetDate:
          type: string
      required:
        - startTime
        - endTime
        - targetHours
        - includeWeekends
        - targetDate

    QuizResult:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        score:
          type: integer
        total:
          type: integer
        percentage:
          type: string
        subject:
          type: string
          nullable: true
        createdAt:
          type: string
          nullable: true
      required:
        - id
        - userId
        - score
        - total
        - percentage

    SaveQuizResultRequest:
      type: object
      properties:
        userId:
          type: string
        score:
          type: integer
        total:
          type: integer
        percentage:
          type: number
        subject:
          type: string
      required:
        - userId
        - score
        - total
        - percentage

    StudySession:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        subjectName:
          type: string
        subjectCode:
          type: string
          nullable: true
        hours:
          type: string
        sessionDate:
          type: string
        completed:
          type: boolean
          nullable: true
        topic:
          type: string
          nullable: true
        createdAt:
          type: string
          nullable: true
      required:
        - id
        - userId
        - subjectName
        - hours
        - sessionDate

    SaveStudySessionRequest:
      type: object
      properties:
        userId:
          type: string
        subjectName:
          type: string
        subjectCode:
          type: string
        hours:
          type: number
        sessionDate:
          type: string
        completed:
          type: boolean
        topic:
          type: string
      required:
        - userId
        - subjectName
        - hours
        - sessionDate