# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Study Planner API
 * OpenAPI spec version: 0.2.0
 */

export interface Profile {
  id: string;
  supabaseId?: string | null;
  name: string;
  email: string;
  college?: string | null;
  createdAt?: string | null;
  updatedAt?: string | null;
}