# lib/api-zod/src/generated/types/saveQuizResultRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Study Planner API
 * OpenAPI spec version: 0.2.0
 */

export interface SaveQuizResultRequest {
  userId: string;
  score: number;
  total: number;
  percentage: number;
  subject?: string;
}