# lib/api-zod/src/generated/types/saveStudySessionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Study Planner API
 * OpenAPI spec version: 0.2.0
 */

export interface SaveStudySessionRequest {
  userId: string;
  subjectName: string;
  subjectCode?: string;
  hours: number;
  sessionDate: string;
  completed?: boolean;
  topic?: string;
}