# lib/api-zod/src/generated/types/studyConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Study Planner API
 * OpenAPI spec version: 0.2.0
 */
import type { StudySettings } from "./studySettings";
import type { Subject } from "./subject";

export interface StudyConfig {
  id: string;
  userId: string;
  courseType: string;
  branch: string;
  year: string;
  semester: string;
  subjects?: Subject[];
  studySettings?: StudySettings;
  createdAt?: string | null;
  updatedAt?: string | null;
}