# lib/api-zod/src/generated/types/studySession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Study Planner API
 * OpenAPI spec version: 0.2.0
 */

export interface StudySession {
  id: string;
  userId: string;
  subjectName: string;
  subjectCode?: string | null;
  hours: string;
  sessionDate: string;
  completed?: boolean | null;
  topic?: string | null;
  createdAt?: string | null;
}