# lib/api-zod/src/generated/types/studySettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Study Planner API
 * OpenAPI spec version: 0.2.0
 */

export interface StudySettings {
  startTime: string;
  endTime: string;
  targetHours: number;
  includeWeekends: boolean;
  targetDate: string;
}