# lib/api-zod/src/generated/types/subject.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Study Planner API
 * OpenAPI spec version: 0.2.0
 */

export interface Subject {
  code: string;
  name: string;
  hours: number;
  difficulty: string;
  description: string;
  selected?: boolean | null;
}