# lib/api-zod/src/generated/types/upsertProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Study Planner API
 * OpenAPI spec version: 0.2.0
 */

export interface UpsertProfileRequest {
  userId: string;
  supabaseId?: string;
  name: string;
  email: string;
  college?: string;
}